forgot to change Cython class numEnvs to be a uint8 to match init prototype, caused silent overflow

+++ cy_impulse_wars.pyx
@@ -77,7 +77,7 @@ def obsConstants(numDrones: int) -> pufferlib.Namespace:
 
 cdef class CyImpulseWars:
     cdef:
-        uint8_t numEnvs
+        uint16_t numEnvs
         uint8_t numDrones
         bint render
         env* envs
